requirements.txt
Django==1.9.4
Markdown==2.6.6
Pillow==3.2.0
chardet==2.2.1
colorama==0.2.5
command-not-found==0.3
django-adminplus==0.5
django-axes==1.6.0
django-bootstrap3==7.0.1
django-filter==0.13.0
django-registration==2.0.4
django-rest==0.0.1
djangorestframework==3.3.3
docopt==0.4.0
html5lib==0.999
language-selector==0.1
mysqlclient==1.3.7
pycurl==7.19.3
pygobject==3.12.0
python-apt==0.9.3.5ubuntu2
reportlab==3.3.0
requests==2.2.1
six==1.5.2
ufw==0.34-rc-0ubuntu2
unattended-upgrades==0.1
urllib3==1.7.1
wheel==0.24.0
